backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
requests
apscheduler
python-dotenv
